CCD Lab 1 AST443
This was a collaborative lab undertaken by me and two other lab student partners; Anthony Masella and Trevor Donovan. The purpose of this experiment was to measure and better understand the properties of Telescopic equipment used at the Mt. Stony Brook Observatory. We took both photometric and spectroscopic data sets in hopes to analyze, calibrate, and conceptualize the observatory setup, process, data, as well as the equipment.