Exoplanets (Lab 2) code
Another AST443 collaborative Code package for exoplanet transit detection. Manipulates 300 photometric images, as well as calibration files. We took our experimentation on HAT-P-6 in order to try and confirm the existence of its exoplanet HAT-P-6b. This code is based on an aperture radius of 6 pixels. The research paper write up is provided in a separate latex file. Used files will also be provided separately.